+++ package.json
@@ -20,6 +20,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
+    "react-loader-spinner": "^5.3.4",
     "react-redux": "^8.1.1",
     "react-router-dom": "^6.14.2",
     "redux-persist": "^6.0.0",
