+++ src/components/Loader/loader.styled.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const LoaderContainer = styled.div`
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
